% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularity.R
\name{artist_charts}
\alias{artist_charts}
\title{Returns an artist's number of Billboard Top 100 songs for a specified time period}
\usage{
artist_charts(artist)
}
\arguments{
\item{artist}{artist's name (quoted)}
}
\value{
A dataframe of all the artist's songs that have made the Billboard Top 100
}
\description{
Returns an artist's number of Billboard Top 100 songs for a specified time period
}
