% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{artist_tracks}
\alias{artist_tracks}
\title{Get all of an artist's tracks that are on Spotify}
\usage{
artist_tracks(artist)
}
\arguments{
\item{artist}{Artist's name (quoted)}
}
\value{
A dataframe of an artist's entire Spotify discography
}
\description{
Get all of an artist's tracks that are on Spotify
}
