% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_sys}
\alias{prep_sys}
\title{Prepares your system to get an access token
*run this code in functions trying to access the api}
\usage{
prep_sys()
}
\description{
Prepares your system to get an access token
*run this code in functions trying to access the api
}
