% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{related_artists}
\alias{related_artists}
\title{Returns related artists}
\usage{
related_artists(artist)
}
\arguments{
\item{artist}{Artist's name (quoted)}
}
\value{
A dataframe of related artists
}
\description{
\code{related_artists} communicates with the Spotify API to return 20 artists that are
related to the selected artist.
}
\examples{
related_artists("pitbull")

}
