% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{top_songs}
\alias{top_songs}
\title{Returns an artist's 10 top songs on Spotify}
\usage{
top_songs(artist)
}
\arguments{
\item{artist}{artist's name (quoted)}
}
\description{
Returns an artist's 10 top songs on Spotify
}
