% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio-features.R
\name{artist_audio_features}
\alias{artist_audio_features}
\title{Get the average audio features for an artist}
\usage{
artist_audio_features(artist)
}
\arguments{
\item{artist}{Artist's name (quoted)}
}
\value{
A dataframe of an artist's averages for audio feature metrics
}
\description{
Get the average audio features for an artist
}
