% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{get_artist_id}
\alias{get_artist_id}
\title{Returns an artist's the unique spotify id}
\usage{
get_artist_id(artist)
}
\arguments{
\item{artist}{artist's name (string)}
}
\description{
Returns an artist's the unique spotify id
}
