% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{get_artist_id}
\alias{get_artist_id}
\title{Returns an artist's unique Spotify ID}
\usage{
get_artist_id(artist)
}
\arguments{
\item{artist}{Artist's name (quoted)}
}
\value{
A string of the artist's unique Spotify ID
}
\description{
Returns an artist's unique Spotify ID
}
\examples{
get_artist_id("johnny cash")

}
