% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{get_artist_id}
\alias{get_artist_id}
\title{Returns an artist's unique Spotify ID}
\usage{
get_artist_id(artist)
}
\arguments{
\item{artist}{Artist's name (quoted)}
}
\value{
A string of the artist's unique Spotify ID
}
\description{
\code{get_artist_id} communicates with the Spotify API to retrieve the unique identifier
for the selected artist. This ID is necessary to access information related to a given
artist in the Spotify API.
}
\examples{
get_artist_id("johnny cash")

}
