% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{get_track_id}
\alias{get_track_id}
\title{Returns an individual track's unique Spotify ID}
\usage{
get_track_id(track)
}
\arguments{
\item{track}{Track's name (quoted)}
}
\value{
A string of the track's unique Spotify ID
}
\description{
Returns an individual track's unique Spotify ID
}
\examples{
get_track_id("rich spirit")

}
