% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{get_track_id}
\alias{get_track_id}
\title{Returns unique track ID}
\usage{
get_track_id(track)
}
\arguments{
\item{track}{track's name (quoted)}
}
\value{
The spotify ID
}
\description{
Returns unique track ID
}
