% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{related_artists}
\alias{related_artists}
\title{Returns related artists}
\usage{
related_artists(artist)
}
\arguments{
\item{artist}{artist's name (quoted)}
}
\value{
A dataframe of related artists
}
\description{
Returns related artists
}
